+++ src/codecs/latex/index.ts
@@ -3,9 +3,9 @@
  */
 
 import stencila from '@stencila/schema'
-import { Encode } from '.'
-import * as pandoc from './pandoc'
-import { VFile } from './vfile'
+import { Encode } from '../..'
+import * as pandoc from '../pandoc'
+import { VFile } from '../../vfile'
 
 export const mediaTypes = ['application/x-latex']
