+++ src/index.ts
@@ -1,28 +1,28 @@
 import * as stencila from '@stencila/schema'
 import mime from 'mime'
 import path from 'path'
-import * as csv from './csv'
-import * as demoMagic from './demo-magic'
-import * as docx from './docx'
-import * as gdoc from './gdoc'
-import * as html from './html'
-import * as ipynb from './ipynb'
-import * as jats from './jats'
-import * as json from './json'
-import * as json5 from './json5'
-import * as latex from './latex'
-import * as md from './md'
-import * as ods from './ods'
-import * as odt from './odt'
-import * as pandoc from './pandoc'
-import * as pdf from './pdf'
-import * as rpng from './rpng'
-import * as tdp from './tdp'
-import * as txt from './txt'
+import * as csv from './codecs/csv'
+import * as demoMagic from './codecs/dmagic'
+import * as docx from './codecs/docx'
+import * as gdoc from './codecs/gdoc'
+import * as html from './codecs/html'
+import * as ipynb from './codecs/ipynb'
+import * as jats from './codecs/jats'
+import * as json from './codecs/json'
+import * as json5 from './codecs/json5'
+import * as latex from './codecs/latex'
+import * as md from './codecs/md'
+import * as ods from './codecs/ods'
+import * as odt from './codecs/odt'
+import * as pandoc from './codecs/pandoc'
+import * as pdf from './codecs/pdf'
+import * as rpng from './codecs/rpng'
+import * as tdp from './codecs/tdp'
+import * as txt from './codecs/txt'
+import * as xlsx from './codecs/xlsx'
+import * as xmd from './codecs/xmd'
+import * as yaml from './codecs/yaml'
 import * as vfile from './vfile'
-import * as xlsx from './xlsx'
-import * as xmd from './xmd'
-import * as yaml from './yaml'
 
 export { default as process } from './process'
