+++ tests/jats.test.ts
@@ -1,4 +1,4 @@
-import * as jats from '../src/jats'
+import * as jats from '../src/codecs/jats'
 import articleSimple from './fixtures/article-simple'
 
 test('invertible', async () => {
